+++ app/src/main/java/com/hrb116/waterlogged/common/preferences/Preferences.kt
@@ -1,8 +1,9 @@
-package com.hrb116.waterlogged.common
+package com.hrb116.waterlogged.common.preferences
 
 enum class Preferences(val key: String) {
     WATER("water.water"),
     WATER_GOAL("water.water_goal"),
     WATER_GOAL_PROGRESS("water.water_goal_progress"),
-    WATER_UNIT("water.water_unit")
+    WATER_UNIT("user.water_unit"),
+    USER_NAME("user.name")
 }
